refactor: make aliase path to files
