+++ harry_potter/src/App.tsx
@@ -1,9 +1,9 @@
 import './App.css';
 import { DataAppProvider } from './context/dataAppProvider';
-import { MainPage } from './components/mainPage';
+import { MainPage } from './pages/mainPage/mainPage';
 import { Route, Routes } from 'react-router';
-import { DetailedCard } from './components/detailedCard';
-import { NotFoundPage } from './components/notFoundPage';
+import { DetailedCard } from './pages/detailed/detailedCard';
+import { NotFoundPage } from './pages/notFound/notFoundPage';
 
 function App() {
   return (
