+++ harry_potter/src/assets/assetsExport.ts
@@ -1,0 +1,2 @@
+import MissCharacter from '../assets/icons/missing_character.svg';
+export { MissCharacter };
