+++ harry_potter/src/components/Card/index.ts
@@ -1,0 +1,1 @@
+export { Card } from './card.tsx';
