+++ harry_potter/src/components/ErrorBoundary/index.ts
@@ -1,0 +1,1 @@
+export { ErrorBoundary } from './errorBoundary.tsx';
