+++ harry_potter/src/components/ErrorModal/index.ts
@@ -1,0 +1,1 @@
+export { ErrorModal } from './errorModal.tsx';
