+++ harry_potter/src/components/Footer/footer.tsx
