+++ harry_potter/src/components/Header/index.ts
@@ -1,0 +1,1 @@
+export { Header } from './header.tsx';
