+++ harry_potter/src/components/NotFound/index.ts
@@ -1,0 +1,1 @@
+export { NotFoundPage } from './notFoundPage.tsx';
