+++ harry_potter/src/components/index.ts
@@ -1,0 +1,6 @@
+export { Card } from './Card';
+export { ErrorBoundary } from './ErrorBoundary';
+export { ErrorModal } from './ErrorModal';
+export { Footer } from './Footer';
+export { Header } from './Header';
+export { NotFoundPage } from './NotFound';
