+++ harry_potter/src/main.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { BrowserRouter } from 'react-router';
 import './index.css';
 import App from './App.tsx';
-import { ErrorBoundary } from './components/errorBoundary.tsx';
+import { ErrorBoundary } from './components/ErrorBoundary/errorBoundary.tsx';
 
 const rootElement = document.getElementById('root');
