+++ harry_potter/src/pages/notFound/notFoundPage.tsx
@@ -1,0 +1,19 @@
+import { useNavigate } from 'react-router';
+
+export function NotFoundPage() {
+  const navigate = useNavigate();
+  return (
+    <>
+      <div className="notfound-container">
+        <h1 className="notfound-title">Page not found</h1>
+        <button
+          className="notfound-button"
+          type="button"
+          onClick={() => navigate('/')}
+        >
+          Main page
+        </button>
+      </div>
+    </>
+  );
+}
