+++ harry_potter/tsconfig.app.json
@@ -20,7 +20,12 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+    /*Path*/
+    "baseUrl": ".",
+    "paths": {
+      "$/*": ["src/*"]
+    }
   },
   "include": ["src"]
 }
